requirements.txt
asgiref==3.3.0
astroid==2.4.2
attrs==20.3.0
awsebcli==3.19.3
bcrypt==3.2.0
blessed==1.17.12
botocore==1.19.56
cached-property==1.5.2
cement==2.8.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cloudinary==1.23.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
dj-database-url==0.5.0
dj-rest-auth==2.1.1
Django==3.1.3
django-allauth==0.42.0
django-ckeditor==6.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.5.0
django-filter==2.4.0
django-grappelli==2.14.2
django-js-asset==1.2.2
django-nested-admin==3.3.3
django-s3direct==1.1.8
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
docker==4.4.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.1
oauthlib==3.1.0
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
pycryptodome==3.9.9
pycryptodomex==3.9.9
PyJWT==1.7.1
pylint==2.6.0
PyNaCl==1.4.0
pyotp==2.6.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
semantic-version==2.5.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.57.0
whitenoise==5.2.0
wrapt==1.12.1
